Move check_and_normalize_axes to Utilies section in basic.py
